fix a bug in uci::Info::update()
Move::to_string() requires a Position object. This is due to !Chess960 case,
where moves like e1h1 must be transformed into e1g1. But the Position must be
updated by playing the moves.

An assert fired in file_of(_tsq), where _tsq was out of bounds. This happens,
for example, with a white King on a1, then playing moves like pv[0] = a1b1, ...,
pv[2] = b1a1. Now pv[2], interpreted with the initial position object, will be
understood as a castling move, because it captures an own piece (Ka1). And the
above transformation leads to _tsq = from - 2 = 1 - 2 = -1: out of bounds!