Avoid auto return type in elapsed()
This is not so good for API, because it does not clearly document what it is.
Plus it requires a more recent compiler, and we prefer to stay C++11 compatible,
with no dependancy on C++14.

Whikt at it, make start a private member of Clock, and put the implementation in
types.cc, to present only a clean API in types.h.