Introduce attacks[][NB_PIECE]
Contains attacks for pieces only (ie. excluding KING and PAWN).

For this, it became necessary to declare the array attacks[][NB_PIECE+1]. This
caused a compile error, because the "+" is a user defined operator, which is not
a constexpr. So we make it a constexpr, and may as well use it in similar places
in the code, where it makes things more self documenting.