Lockless hashing
No more SMP crashes. However, this is not a real fix. All it does is make
collisions so rare that we don't experience crashes in practice. But when
collisions do occur, they can still cause crashes, which is a bug that has yet
to be understood and fixed.

Scaling is pretty good on 8 HT threads (4 cores):

8 threads vs 1 thread: 686 - 93 - 421  [0.747] 1200
ELO difference: 188.16 +/- 16.55

Testing conditions:

./cutechess-cli \
  -each proto=uci tc=4+0.04 option.Hash=16 \
  -openings file='./book5.epd' format=epd -repeat \
  -engine cmd=$1 name="8 threads" option.Threads=8 \
  -engine cmd=$2 name="1 thread" \
  -pgnout $3 -recover \
  -concurrency 1 -rounds 1200 -ratinginterval 25 \
  -draw movenumber=40 movecount=8 score=20 \
  -resign movecount=3 score=500 \