Formatting
Use *.h instead of *.inc extension for recurse. This is a an edge case. On the
one hand, it's a header, because it's not a translation unit in itself. On the
other hand, it contains code only, rather than declarations, as is traditional
for headers. Choose a standard extension, to make usual C/C++ toolchain
recognize it.

Also fix spacing around operators.

Progress update:

demolito vs discocheck_5.2.1: 2120 - 4830 - 3050  [0.364] 10000
ELO difference: -96.57 +/- 5.79

Testing conditions:

./cutechess-cli \
  -each proto=uci tc=8+0.08 option.Hash=8 \
  -openings file='./book5.epd' format=epd -repeat \
  -engine cmd=$1 name=$1 \
  -engine cmd=$2 name=$2 \
  -pgnout $3 -recover \
  -concurrency 7 -rounds 10000 -ratinginterval 25 \
  -draw movenumber=40 movecount=8 score=20 \
  -resign movecount=3 score=500