Persistent PawnHash
This is a non-functional speedup. However, it's difficult to measure it with
test_search() run(s). This is because positions are not so related to each
other, whereas in real games, the next search position generally has (almost)
the same pawn structure as the previous one. So we get a lot more PawnHash hits
from previous search(es). Hence real game testing, but with the shortest tc
possible, to increase the sensitivity of the measure.

SPRT(0,4), tc=2+0.02, hash=2:

persistent vs master: 6371 - 6098 - 8590  [0.506] 21059
ELO difference: 4.50 +/- 3.61
SPRT: llr 2.95, lbound -2.94, ubound 2.94 - H1 was accepted