Allow eval before go
This resolves a segfault, triggered for example by:

$ ucinewgame
$ position startpos
$ eval

What happens is that thisWorker is NULL, which segfaults in pawn evaluation when
trying to use thisWorker->pawnHash.

Although not a searching thread per se, the master thread still needs a valid
thisWorker = &Worker[0] to be able to call evaluate(). Number of threads can
change, but what remains constant for the life of the UCI code is that we have
at least 1 element in Workers[] array, and that its pawnHash is either empty or
holds valid data.