Use C11 threads in search
Something's wrong with these stupid min/max macros. Although defined in types.h,
they are not usable in uci.cc. This is really annoying, and it's unclear which
of the headers is screwing that up. include <thread> was hiding the problem, and
causing me to use the std::min() and std::max() from <algorithm>, which I cannot
use in C. For the moment, just do without them.