Cosmetic touch
Don't use hex notation, as there's no need for it. We need a power of 2, for
fast modulo, and something big enough in practice, that's all.

More importantly, remove the misleading 2 * 64 * 64. This leads to the incorrect
interpretation that we need 2 (colors) * 64 (from) * 64 (to). But this is not a
correct way to enumerate quite moves. Instead one needs to count all the unique
quadruplets (color, piece, from, to). This is double the number of unique
triplets (piece, from, to), which is 336(N) + 560(B) + 896(R) + 1456(Q) + 420(K)
+ 56(P) = 3727. So the total is 7448, rounded to 8192.