Statically compile with musl
It's incredible the amount of bloat in glibc. Using gcc to do a static compile
against glibc and musl, the difference is staggering:
- musl: 78kB
- glibc: 1026kB

So nice to have a GNU free exe. Bloatless and totally portable (dynamically
compiling is never really portable to all linux distro).

For now keep "threads.h", even though it could be replaced by <threads.h> from
musl. The thing is that C11 <threads.h> and musl are not easily available
outside of Linux (eg. Android or MacOSX).