Revert musl
Somehow musl crashes randomly, when compiling with gcc 6.3, but not 5.4. The
culprit could be musl, but could also be gcc, or even my code (that somehow
relies on a more robust behavior of glibc that isn't guaranteed by the
standard). I don't have time to investigate libc crashes, so revert musl for
now.