Enable reliable speed measures on Windows
Writing to a terminal is very slow, and we write a lot:

$ ./master search 12 1 | wc -l
1342

On Linux, the way to solve this and not pollute speed measures with the terminal
displaying speed, is for example:

$ ./master search 12 1 | tail -2
kn/s: 1874
total = 11680423

This is nice, because it exactly reproduces real conditions. Here stdout is a
pipe, and tail is reading the other end of the pipe, just like a GUI would.

The problem is that it's not possible on Windows, as there is no 'tail' command,
or anything that could be used to replace it (eg. 'grep'). So, instead, write to
stderr the last 2 lines, so that speed can be measured on Windows like this:

$ demolito.exe search 12 1 1 > nul

I don't have Windows to test it, but assume it works.

While there, remove the pos_print(). The idea is to reproduce real conditions as
close as possible (no more, no less, same amount of writing to stdout).