Fix PV termination
Bug spotted using Valgrind. In (rare) cases, the pv[0]=0 is not being done,
because pvNode=false, even though we are actually in a PV node, that just
happens to have satisfy alpha+1 == beta.

Another possibility would be to fix the definition of pvNode, but that means
passing it recursively, which I find inelegant. Besides, these "fake PV nodes"
with alpha+1 == beta are somewhat suspicious, and prone to making wrong
assumptions. I like to be able to assume true PV nodes with a non empty
[alpha,beta[ interval.

This is actually a mesurable slowdown:

Score of demolito vs before: 10712 - 10964 - 18926  [0.497] 40602
Started game 40609 of 63052 (./Demolito/demolito vs ./Demolito/before)

But we cannot afford to keep bugs like that, so let's commit anyway.