Speedup QS non-PV nodes (13867731)
Gains back some of the speed lost in the previous patch. Measured 1% speed-up on
the median of 5 runs of:

$ ./demolito search 12 1 > /dev/null

Avoid the expensive pv[0]=0 at non-PV nodes. However, to be able to do that
safely, one has to have a 100% bullet proof definition of PV nodes, which means
passing it recursively.

Due to rare cases where PV nodes appear with a null window, the node count is
changed. For good measure, I've verified that the bench is restored by adding:

if (pvNode && alpha + 1 == beta)
    pvNode = false;

So the only functional difference comes from these degenerated PV nodes.