Correct SEE for HT move (13625873)
This fixes a bug, where qsearch() returned scores out of bounds in the case
staticEval + see. This bug was exposed by removing INF and instead using MATE
for search, and INT_MIN/MAX for sort (to better document the code).

Previous the HT move got see = INF - SEPARATION. This was unnoticed, because the
value staticEval + see was still within bounds. At least often enough not to
trigger an out of bounds score in demolito search 12 1.

This seems elo neutral, though could be a slight slowdown. Anyway, stability
cannot be compromised for the sake of micro optimisation, so this bug is worth
fixing, even if it cots a tiny bit of elo.

SPRT(-4,0), tc=4+0.04, hash=4:
demolito vs previous: 2906 - 2878 - 5087  [0.501] 10871