Better manage hash_resize() and smp_resize()
Firstly, we don't want to rely on ucinewgame. If the user forgets this, the
program cannot crash, that is violating the UCI protocol (although a well
designed UI won't make that mistake, we have to heed the protocol, which says
ucinewgame can't be relied on).

That means the responsability of allocating the hash table firmly belongs to
main(), which is also coherent with where we free() it at program exit.

Same logic should apply to smp_resize().

Also, rename resize to prepare, because we do resize + clear, not just resize.

Also, ensure that uciHash and WorkersCount are always consistent, not only in
UCI mode, but also in bench mode (perft or search).