Shrink history tables
A very small speed-up. Both with Hash=2 and Hash=16 (to ensure it's not just a
non scalable cache effect of Hash=2).

Hash=2: 0.53% speed-up
Hash=16: 0.31% speed-up

As long as it's not a slowdown, it's good to reduce the memory footprint, so we
can test more things like RefutationHistory, which looks promising.

Progress update:

demolito vs critter_1.6a: 1720 - 5510 - 2770  [0.310] 10000
Elo difference: -138.59 +/- 6.04, LOS: 0.00 %, DrawRatio: 27.7 %

Testing conditions:

./cutechess-cli \
    -each proto='uci' tc=8+0.08 option.Hash=8 option.Threads=1 \
    -engine name=$1 cmd=$1 \
    -engine name=$2 cmd=$2 \
    -openings file='./spsa/book.epd' format='epd' -repeat \
    -draw movenumber=40 movecount=8 score=20 \
    -resign movecount=3 score=500 \
    -concurrency 7 -rounds 5000 -games 2 -ratinginterval 50