Implement infinite correctly
The UCI spec is quite clear. If infinite is specified, the search cannot stop,
even if the max depth is completed. However, a simple "go" without infinite,
doesn't have this requirement, so in this case it's ok to return a bestmove.

This will also be useful for pondering, because go ponder has the same
requirement.