Branchless pawn moves generation (11963786)
Except for promotions, which are rare enough, that we can afford branches. By
measuring, using perft, it is not clear at all whether this is a speed-up or
not. It is within the noise level. But there is an undeniable elo gain, as shown
below. Perhaps it's the move ordering effect, where putting pawn captures first,
for equal sort score, means we try first to search a sub-tree with fewer pieces
on the board, and therefore a lower branching factor. Or perhaps a small speed
gain. Or a combination of both. Anyway, commit this patch even though I do not
like ugly code. There could be something there.

SPRT(0,4), tc=2+0.02, hash=2:
gen vs demolito: 5128 - 4881 - 7752  [0.507] 17761
SPRT: llr 2.96 (100.5%), lbound -2.94, ubound 2.94 - H1 was accepted

SPRT(0,4), tc=4+0.04, hash=4:
gen vs demolito: 1524 - 1360 - 2728  [0.515] 5612
SPRT: llr 2.95 (100.3%), lbound -2.94, ubound 2.94 - H1 was accepted