Portable aligned_alloc()
Yet another portability shitshow...

In theory, we should have:
* aligned_alloc() from C11
* posix_memalign() from POSIX
* _aligned_malloc() from MSDN

In practice...
* aligned_alloc() is only available for Linux and Anrdoid cross compiler.
* posix_memalign() is broken on MacOSX. I find this hard to believe, but I don't
    have a Mac to verify, so I have to trust @SpillTheTea on this one.
* _aligned_malloc() is silently broken for Windows. The compiler doesn't
    complain, but fails to link it statically, and the compile only works with a
    bloated 3MB DLL (confirmed by @erbsenzaehler).

So reinvent the wheel and write our own aligned malloc/free. This is more
general and cleaner than the SF, because only a single pointer is needed,
instead of 2. I might reuse this for PawnHash, if I decided to dynamically
allocated it per thread, to make it NUMA friendly.