move_is_legal() -> gen_is_legal()
This function belongs in the gen unit, because it is intertwined with move
generators. What exactly it checks for, depends on the choices made in move
generators, which are a changing mix of legal and pseudo-legal.

It is not a general purpose legality check, that can be fed random data.