Prevent int16_t overflow (14857959)
This code is extremely dangerous:

*t += 32 * bonus - *t * abs(bonus) / 256;

The problem is that *t is int16_t, and since signed integer overflow is
undefined in C, the compiler can evaluate this code however it wants, leading
to different outcomes, so the overflow effect (eg. wrap around zero on x86)
will happen at different steps. Regardless, the wrap around zero is not what we
want. We want to do the full calc w/o overflow, apply the cap & floor, and only
then store it into 16-bits for reducing memory usage.

This is actually a functional change, albeit an untestable one, since it only
triggers at very high depths.