Make hash probe more robust
There is not undefined behavior yet, but it is too easy to inadvertently add
one. For example, modify hash probing by adding a pre-condition:

if (condition && hash_read(...)) {...}

And if condition is false, then hash_read() is not called, which means we forget
to set he.data = 0.

Since he is owned by search() and qsearch(), invalidating he.data should be done
explicitely there.