Raise Threads limit
Now that I got rid of the selective abort logic, there is no theoretical limit
to the number of threads Demolito can handle. Whether Demolito actually scales
(in elo not nps) beyond 64 threads is another question, and this is completely
untested and untestable.

But TCEC want to break the thread limit, so let's go and break it.