Fix rare crash on HT collisions
Due to a race, or even a collision (single threaded), the he.eval can be wrong.
And so we cannot deduce that we are not in check by trusting this value. And, of
course, doing a null move in check is a disaster, the opponent will capture our
king, and the engine crashes soon after.

Hopefully Demolito is now ready for 176 threads in TCEC without crashes :)