Complete integrity check in finish()
This goes beyond the scope of the previous patch, which was only validating all
the wrong inputs that can come from a FEN string. Here we check all the
invariants of a Position, on every call to finish(), which is called whenever a
position is setup (setup FEN, play move, play null move).

Also improve the piece count conditions. For example, the max for white knights
is 10, but if there are 5 white pawns, then only 3 could have promoted, so the
max is 5.

seal  : 5db854abf054f116
time  : 4305ms
nodes : 11492179
nps   : 2669496