Allow one more depth in endgames for levels
Relative to humans, "weak" levels that are depth bound (level <= 8) are relatively weaker in the endgame than the middle game. An additional depth is a simple way to compensate for that, keeping the same nodes limit in any case (although it is only enforced in time steps of 5ms, so it's almost inactive there in practice).

No functional change when Level=0 (off).

nodes : 8608419